move ReadeAndCreateOrUpdateDeployItem function to test library
